+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ##
+## Added
+- setting Run document macros
+
 ## Changed
 - updated list supported formats, see [ONLYOFFICE/document-formats v3.0.0](https://github.com/ONLYOFFICE/document-formats/releases/tag/v3.0.0)
 - fix CVE-2025-27820